Added and refactored exports of the schema-blocks package, based on the needs of the apply button.
